Adds more to examples and .bat files
Changeing the name of the example files, Adding a new file to that file that adds the basics structure and files for makeing a mod so it can be run in arma.

changeing and adding .bat files, mainly to make them more indapendent so they dont require someone to change much of them to be able to be used by anyone with the repo. as well as makeing them a little better. Mainly makeing it a 1 click done situation