Fixed and got Uniforms working
FUCK uniforms, ok so uniforms require that a solder base and useing the solder name we can create a uniform item, we dont define the texture in the uniform but in the solder, witch means that we need to create a new solder for each Uniform. Have a Bug  That keeps poping up, Required to fix was too define eveything about the hitpoints on the base unit. Hopefully stable now!